third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/delayed-discovery-no-permission-absent-service.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let expected = new DOMException('Origin is not allowed to access the ' +
                                  'service. Tip: Add the service UUID to ' +
                                  '\'optionalServices\' in requestDevice() ' +
                                  'options. https://goo.gl/HxfxSQ',
                                  'SecurityError');
  return setBluetoothFakeAdapter('DelayedServicesDiscoveryAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
      assert_promise_rejects_with_message(
        gattServer.getPrimaryService(glucose.alias), expected),
      assert_promise_rejects_with_message(
        gattServer.getPrimaryService(glucose.name), expected),
      assert_promise_rejects_with_message(
        gattServer.getPrimaryService(glucose.uuid), expected)]));
}, 'Delayed services discovered, request for absent service without ' +
   'permission. Reject with SecurityError.');
</script>
